Refactor: move styling of type title into its CSS and make it not h2
